Start using NumPy for speed
NumPy is a library to make it easy to write high-performance data manipulation and number
crunching code in Python. This looks like the key to making this run fast on the Raspberry Pi.

Profiling showed that most of our CPU time was being spent in the opc_client, so that was
rewritten using NumPy. None of the effects take advantage of NumPy features yet, but we could
choose to do so if we're hitting performance bottlenecks in individual effects later.
Otherwise, effects can keep treating them like normal arrays.